// src/lib/api/generated/model/campaign.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { DeliveryChannelEnum } from './deliveryChannelEnum';
import type { CampaignStatusEnum } from './campaignStatusEnum';

export interface Campaign {
  readonly id: string;
  readonly tenant: string;
  readonly created_by: string;
  form: string;
  delivery_channel?: DeliveryChannelEnum;
  /** @nullable */
  email_list?: string | null;
  /** @maxLength 200 */
  webhook_url?: string;
  /** @maxLength 255 */
  name: string;
  description?: string;
  status?: CampaignStatusEnum;
  /** @nullable */
  start_date?: string | null;
  /** @nullable */
  start_time?: string | null;
  /** @maxLength 64 */
  timezone?: string;
  /** @nullable */
  readonly max_recipients: number | null;
  readonly recipients_sent: number;
  readonly recipients_responded: number;
  readonly created_at: string;
  readonly updated_at: string;
}