// src/lib/api/generated/model/metricResult.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Django Continuous Integration API
 * API documentation for the Django Continuous Integration project with tenant-aware access rules.
 * OpenAPI spec version: 1.0.0
 */
import type { MetricTypeEnum } from './metricTypeEnum';

export interface MetricResult {
  readonly id: string;
  readonly metric_type: MetricTypeEnum;
  /** @pattern ^-?\d{0,4}(?:\.\d{0,2})?$ */
  readonly value: string;
  readonly period: string;
  readonly form_id: string;
  /** @nullable */
  readonly campaign_id: string | null;
  readonly created_at: string;
  readonly updated_at: string;
}